requirements.txt
langchain 
langchain-community 
langchain-text-splitters 
langchain-openai
pandas
pypdf
chromadb
langchain-chroma
rapidfuzz
olefile
pyhwp  # HWP file parser for Korean documents
python-docx  # DOCX file parser
pyhwp
rank_bm25
pyyaml
tqdm
numpy

opencv-python 
pytesseract

qdrant-client
langchain-qdrant
rank-bm25
flashrank
fastapi
uvicorn

# Evaluation
rouge-score
nltk

# Environment variables
python-dotenv